Implement TileMatrix#toString, overload TileMatrix constructor, implement Tile#isIdenticalTo
TileMatrix#toString is basically a "display" to be "printed."
The TileMatrix constructor is now overloaded. The original constructor with 'int rows' and 'int columns' which makes a matrix full of empty tiles. The new constructor simply takes a Tile[][] and assigns its matrix to that directly.
Tiles are equal if they are the exact same (Tile this == Tile that). Tiles are *identical* if they share equal contents (String this.equals(String that) and the same color code (Color this.ansiCode.equals(Color that.ansiCode))).